# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface ListingPrompt {
  /**
     * @minLength 3
     * @maxLength 2000
     */
  raw: string;
  /**
     * @minLength 1
     * @maxLength 80
     */
  category: string;
}

export type GeneratedListingCurveType = typeof GeneratedListingCurveType[keyof typeof GeneratedListingCurveType];


export const GeneratedListingCurveType = {
  steep_4hr: 'steep_4hr',
  linear_24hr: 'linear_24hr',
  flat_2week: 'flat_2week',
} as const;

export interface GeneratedListing {
  title: string;
  description: string;
  quantity: string;
  /**
     * @minimum 0
     * @maximum 100
     */
  estimated_condition: number;
  curve_type: GeneratedListingCurveType;
  /** @exclusiveMinimum 0 */
  starting_price_inr: number;
  curve_reason: string;
}

export interface HealthStatus {
  status: string;
}