# lib/api-zod/src/generated/types/generatedListing.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { GeneratedListingCurveType } from './generatedListingCurveType';

export interface GeneratedListing {
  title: string;
  description: string;
  quantity: string;
  /**
     * @minimum 0
     * @maximum 100
     */
  estimated_condition: number;
  curve_type: GeneratedListingCurveType;
  /** @exclusiveMinimum 0 */
  starting_price_inr: number;
  curve_reason: string;
}