# lib/api-zod/src/generated/types/listingPrompt.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ListingPrompt {
  /**
     * @minLength 3
     * @maxLength 2000
     */
  raw: string;
  /**
     * @minLength 1
     * @maxLength 80
     */
  category: string;
}